Move anf stuff into right place
